Separated CameraComponent and DebugMovementComponent
