+++ crates/engine/examples/minimal/main.rs
@@ -10,7 +10,10 @@ use vulkan_engine::{
         window::{self, Dimensions},
     },
     scene::{
-        component::{camera_component::CameraComponent, renderer::RendererComponent},
+        component::{
+            camera_component::CameraComponent, debug_movement_component::DebugMovementComponent,
+            renderer::RendererComponent,
+        },
         material::MaterialPipeline,
         model::{mesh::Mesh, Model},
         transform::Transform,
@@ -137,6 +140,7 @@ fn setup(engine: &mut Engine) {
         },
     );
     main_cam.new_component::<CameraComponent>();
+    main_cam.new_component::<DebugMovementComponent>();
 
     scene.load();
