+++ crates/engine/examples/textured_material/main.rs
@@ -10,7 +10,10 @@ use vulkan_engine::{
         window::{self, Dimensions},
     },
     scene::{
-        component::{camera_component::CameraComponent, renderer::RendererComponent},
+        component::{
+            camera_component::CameraComponent, debug_movement_component::DebugMovementComponent,
+            renderer::RendererComponent,
+        },
         light::*,
         material::MaterialPipeline,
         model::{mesh::Mesh, Model},
@@ -167,6 +170,7 @@ fn setup(engine: &mut Engine) {
         },
     );
     main_cam.new_component::<CameraComponent>();
+    main_cam.new_component::<DebugMovementComponent>();
 
     scene.load();
